Updated vue2 and vue3 sample
