+++ README.md
@@ -1,2 +1,2 @@
-# View and edit Word document in Vue
-View and edit Word document in Vue application without Microsoft Word or Office interop dependencies.
+# View and edit Word document in Vue2 and Vue3
+View and edit Word document in Vue2 and Vue3 application without Microsoft Word or Office interop dependencies.
