+++ Vue2/.editorconfig
