+++ Vue2/README.md
@@ -1,0 +1,2 @@
+# View and edit Word document in Vue2
+View and edit Word document in Vue2 application without Microsoft Word or Office interop dependencies.
