+++ Vue3/README.md
@@ -1,0 +1,12 @@
+# View and edit Word document in Vue3
+View and edit Word document in Vue3 application without Microsoft Word or Office interop dependencies.
+
+## Project setup
+```
+npm install
+```
+
+### To run
+```
+npm run serve
+```
