+++ Vue3/public/index.html
@@ -1,0 +1,16 @@
+<div>
+  <div ref="de_titlebar" id="documenteditor_titlebar" class="e-de-ctn-title">
+    <div v-on:keydown="titleBarKeydownEvent" v-on:click="titleBarClickEvent" class="single-line"
+      id="documenteditor_title_contentEditor" title="Document Name. Click or tap to rename this document."
+      contenteditable="false">
+      <label v-on:blur="titleBarBlurEvent" id="documenteditor_title_name" :style="titileStyle">{{documentName}}</label>
+    </div>
+    <ejs-button id="de-print" :style="iconStyle" :iconCss="printIconCss" v-on:click="printBtnClick"
+      title="Print this document (Ctrl+P).">Print</ejs-button>
+    <ejs-dropdownbutton ref="de-export" :style="iconStyle" :items="exportItems" :iconCss="exportIconCss"
+      cssClass="e-caret-hide" content="Download" v-bind:select="onExport" :open="openExportDropDown"
+      title="Download this document."></ejs-dropdownbutton>
+  </div>
+  <ejs-documenteditorcontainer id='container' ref="doceditcontainer" :enableToolbar='true' height="590px"
+    :serviceUrl='serviceUrl'></ejs-documenteditorcontainer>
+</div>
